refactor(feat-3298): move GraphService and MockGraphService to Adapters.Microsoft365
Moves concrete Graph service implementations out of the Application layer
into the Microsoft365 adapter project where they architecturally belong.
IGraphService interface remains in Application. DI registrations migrated
from UserManagementModule to Microsoft365AdapterServiceCollectionExtensions.
