+++ backend/src/Adapters/Anela.Heblo.Adapters.Microsoft365/Microsoft365AdapterServiceCollectionExtensions.cs
@@ -1,6 +1,8 @@
 using Anela.Heblo.Adapters.Microsoft365.Photobank;
+using Anela.Heblo.Adapters.Microsoft365.UserManagement;
 using Anela.Heblo.Application.Features.Marketing.Services;
 using Anela.Heblo.Application.Features.Photobank.Services;
+using Anela.Heblo.Application.Features.UserManagement.Services;
 using Anela.Heblo.Domain.Features.Configuration;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -13,10 +15,15 @@ public static class Microsoft365AdapterServiceCollectionExtensions
         this IServiceCollection services,
         IConfiguration configuration)
     {
-        var useMockAuth = configuration.GetValue<bool>("UseMockAuth", false);
+        var useMockAuth = configuration.GetValue<bool>(ConfigurationConstants.USE_MOCK_AUTH, false);
         var bypassJwt = configuration.GetValue<bool>(ConfigurationConstants.BYPASS_JWT_VALIDATION, false);
 
-        if (!useMockAuth && !bypassJwt)
+        if (useMockAuth || bypassJwt)
+        {
+            // Register mock GraphService for mock / dev authentication
+            services.AddScoped<IGraphService, MockGraphService>();
+        }
+        else
         {
             services.AddScoped<IOutlookCalendarSync, OutlookCalendarSyncService>();
             services.AddHttpClient("MicrosoftGraph", _ => { })
@@ -25,6 +32,11 @@ public static class Microsoft365AdapterServiceCollectionExtensions
                     AllowAutoRedirect = true,
                 });
             services.AddScoped<IPhotobankGraphService, PhotobankGraphService>();
+
+            // Register real GraphService for production authentication
+            // Note: GraphServiceClient must be registered in the API layer with proper authentication
+            // through Microsoft.Identity.Web's AddMicrosoftGraph() method
+            services.AddScoped<IGraphService, GraphService>();
         }
 
         return services;
